Move molecule hypothesis strategies in their own module
